feat(task6): refactor code
